# Conflicts:
#	.idea/dataSources/7ce76d44-e311-4ffb-808a-6f6317dc809c/_metadata_/metadata
#	.idea/workspace.xml
#	src/main/java/com/authentification/userDAO/UserDAOImpl.java
#	src/main/java/com/test/TestClass.java
#	src/main/java/servlets/Logout.java
#	src/main/java/servlets/RegisterServlet.java
#	src/main/java/servlets/admin/AddUser.java
#	src/main/java/servlets/admin/MainAdmin.java
#	src/main/java/servlets/staff/ConstructorModel.java
#	src/main/webapp/WEB-INF/web.xml
#	src/main/webapp/admin/admin.jsp
#	src/main/webapp/img/user.png
#	src/main/webapp/staff/StaffModel.jsp
#	target/MySiteArt-1.0-SNAPSHOT.war
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/com/authentification/userDAO/UserDAOImpl.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/com/test/TestClass.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/RegisterServlet.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/admin/AddUser.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstructorBody.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstructorColor.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstructorModel.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstructorResult.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstructorTransmission.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/classes/servlets/user/ConstuctorEngine.class
#	target/MySiteArt-1.0-SNAPSHOT/WEB-INF/web.xml
#	target/MySiteArt-1.0-SNAPSHOT/admin/admin.jsp
#	target/MySiteArt-1.0-SNAPSHOT/img/user.png
#	target/MySiteArt-1.0-SNAPSHOT/staff/StaffModel.jsp
#	target/classes/com/authentification/userDAO/UserDAOImpl.class
#	target/classes/com/test/TestClass.class
#	target/classes/servlets/RegisterServlet.class
#	target/classes/servlets/admin/AddUser.class
#	target/classes/servlets/user/ConstructorBody.class
#	target/classes/servlets/user/ConstructorColor.class
#	target/classes/servlets/user/ConstructorModel.class
#	target/classes/servlets/user/ConstructorResult.class
#	target/classes/servlets/user/ConstructorTransmission.class
#	target/classes/servlets/user/ConstuctorEngine.class
#	target/maven-status/maven-compiler-plugin/compile/default-compile/createdFiles.lst
#	target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst

+++ src/main/java/com/carDAO/ColorDAO.java
@@ -58,7 +58,7 @@ public class ColorDAO implements DAO<Color> {
     }
 
     @Override
-    public void update(Long l, Color color) throws SQLException {
+    public void update(Integer l, Color color) throws SQLException {
         Session session = null;
         try{
             session = HibernateUtil.getSessionFactory().openSession();
@@ -75,7 +75,7 @@ public class ColorDAO implements DAO<Color> {
     }
 
     @Override
-    public Color getById(Long l) {
+    public Color getById(Integer l) {
         Session session = null;
         Color color = null;
         try{
@@ -109,7 +109,7 @@ public class ColorDAO implements DAO<Color> {
     }
 
     @Override
-    public void delete(Long l) {
+    public void delete(Integer l) {
         Session session = null;
         Color color =null;
         try{
