+++ src/main/java/com/carEntity/Transmission.java
@@ -13,9 +13,9 @@ import java.io.Serializable;
 public class Transmission implements CarParts, Serializable {
 
 
-    private static final long serialVersionUID = 1L;
+    private static final Long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
     private String transmissionName;
     private int numberOfSpeed;
 
@@ -29,11 +29,11 @@ public class Transmission implements CarParts, Serializable {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name = "idtransmission",nullable = false, unique = true)
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
